Fixed employee rate of pay so that it is of type double rather than an int. Fixed it so that any monetary fields are formatted to 2 decimal places and preceded by a pound sign.

+++ app/src/main/java/com/endorphinapps/kemikal/queenofclean/Entities/Employee.java
@@ -17,7 +17,7 @@ public class Employee {
     private String mobileNumber;
     private String homeNumber;
     private String emailAddress;
-    private int rateOfPay;
+    private double rateOfPay;
 
     public Employee() {
     }
@@ -27,7 +27,7 @@ public class Employee {
                     String addressLine2,  String town, String city,
                     String postcode, String mobileNumber,
                     String homeNumber, String emailAddress,
-                    int rateOfPay) {
+                    double rateOfPay) {
         this.employeeId = employeeId;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -130,11 +130,11 @@ public class Employee {
         this.emailAddress = emailAddress;
     }
 
-    public int getRateOfPay() {
+    public double getRateOfPay() {
         return rateOfPay;
     }
 
-    public void setRateOfPay(int rateOfPay) {
+    public void setRateOfPay(double rateOfPay) {
         this.rateOfPay = rateOfPay;
     }
